fix(Player): 플레이어가 버스트되었을 때 플레이어가 무조건 패배하도록 로직 변경
- compareWith 로직을 Participant 에서 Player 로 변경
- 테스트 코드 변경
