+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -1,8 +1,6 @@
 package blackjack.domain.participant;
 
 import static blackjack.domain.fixture.CardRepository.CLOVER10;
-import static blackjack.domain.fixture.CardRepository.CLOVER2;
-import static blackjack.domain.fixture.CardRepository.CLOVER3;
 import static blackjack.domain.fixture.CardRepository.CLOVER4;
 import static blackjack.domain.fixture.CardRepository.CLOVER5;
 import static blackjack.domain.fixture.CardRepository.CLOVER6;
@@ -13,8 +11,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Hand;
-import blackjack.domain.game.ResultType;
-import blackjack.domain.money.Money;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -73,50 +69,6 @@ public class DealerTest {
         boolean actual = dealer.canReceive();
 
         assertThat(actual).isFalse();
-    }
-
-    // TODO: Nested 사용하여 정리
-
-    @DisplayName("compareWith 메소드는 딜러 자신보다 패가 나쁜 Participant 를 전달받으면 ResultType.WIN 를 반환한다.")
-    @Test
-    void compareWith_returnsResultTypeWin() {
-        // given
-        Player winPlayer = Player.of("hudi", Hand.of(CLOVER2, CLOVER3), Money.from(10000));
-
-        // when
-        ResultType actual = dealer.compareWith(winPlayer);
-        ResultType expected = ResultType.WIN;
-
-        // then
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    @DisplayName("compareWith 메소드는 딜러 자신보다 패가 좋은 Participant 를 전달받으면 ResultType.LOSE 를 반환한다.")
-    @Test
-    void compareWith_returnsResultTypeLose() {
-        // given
-        Player winPlayer = Player.of("hudi", Hand.of(CLOVER5, CLOVER6), Money.from(10000));
-
-        // when
-        ResultType actual = dealer.compareWith(winPlayer);
-        ResultType expected = ResultType.LOSE;
-
-        // then
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    @DisplayName("compareWith 메소드는 딜러 자신과 대등한 패를 가진 Participant 를 전달받으면 ResultType.DRAW 를 반환한다.")
-    @Test
-    void compareWith_returnsResultTypeDraw() {
-        // given
-        Player winPlayer = Player.of("hudi", Hand.of(CLOVER3, CLOVER6), Money.from(10000));
-
-        // when
-        ResultType actual = dealer.compareWith(winPlayer);
-        ResultType expected = ResultType.DRAW;
-
-        // then
-        assertThat(actual).isEqualTo(expected);
     }
 
     @DisplayName("isBlackjack 은 딜러가 받은 첫 두장의 카드의 합이 21일 경우 true 를 반환한다.")
